filesystem: now trigger a signal when remote file is downloaded
